backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
gunicorn==23.0.0
whitenoise==6.8.2
pydantic==2.10.4
pydantic-settings==2.7.1
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
google-generativeai==0.3.2
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
